+++ playground/supabase-auth/app/billing/page.tsx
@@ -1,7 +1,18 @@
-import { BillingPage } from '@flowglad/nextjs';
+import { BillingPage, PricingTable } from '@flowglad/nextjs';
 import { flowgladServer } from '@/utils/flowglad';
 
 export default async () => {
   const billing = await flowgladServer.getBilling();
-  return <BillingPage billing={billing} />;
+  return (
+    <PricingTable
+      products={billing.catalog.products.map((product) => ({
+        name: product.name,
+        description: product.description,
+        displayFeatures: product.displayFeatures,
+        primaryButtonText: 'Subscribe',
+        secondaryButtonText: 'Learn More',
+        prices: product.prices
+      }))}
+    />
+  );
 };