+++ packages/react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @flowglad/react
 
+## 0.8.9
+
+### Patch Changes
+
+- Relative route check
+- Updated dependencies
+  - @flowglad/shared@0.8.9
+
+## 0.8.8
+
+### Patch Changes
+
+- Export SubscriptionDetails type
+- Updated dependencies
+  - @flowglad/shared@0.8.8
+
 ## 0.8.7
 
 ### Patch Changes